+++ src/commands/FunctionCommands.js
@@ -3,11 +3,11 @@ const vscode = require('vscode')
 const Core = require('../Core')
 const Validator = require('../Validator')
 
-const { VM } = require('vm2')
+const { VM, VMScript } = require('vm2')
 const { IML } = require('@integromat/iml')
 
 class FunctionCommands {
-    static async register(appsProvider, _authorization, _environment) {
+    static async register(appsProvider, _authorization, _environment, _timezone) {
 
         var outputChannel = vscode.window.createOutputChannel("IML tests")
 
@@ -47,23 +47,64 @@ class FunctionCommands {
          */
         vscode.commands.registerCommand('apps-sdk.function.test', async function (context) {
 
-            // If called out of context -> die
-            if (!Core.contextGuard(context)) { return }
-
-            // Set correct URN (if called from function or core or test)
             let urn
-            if (context.supertype === "function") {
-                urn = `${_environment}/app/${Core.getApp(context).name}/${Core.getApp(context).version}/function/${context.name}`
+            let functionName
+
+            // If called out of context -> gather the required information
+            if (context === undefined || context === null) {
+                let crumbs
+                // If a window is open
+                if (!vscode.window.activeTextEditor) {
+                    vscode.window.showErrorMessage("Active text editor not found.")
+                    return
+                }
+                // If an apps file is open
+                if (!vscode.window.activeTextEditor.document.uri.fsPath.split('apps-sdk')[1]) {
+                    vscode.window.showErrorMessage("Opened file doesn't belong to Apps SDK.")
+                    return
+                }
+                // Parse the path
+                crumbs = vscode.window.activeTextEditor.document.uri.fsPath.split('apps-sdk')[1].split('/').reverse()
+                // If crumbs were parsed
+                if (!crumbs) {
+                    vscode.window.showErrorMessage("The path was not parsed successfully.")
+                    return
+                }
+                // If the path hasn't 7 crumbs it's definitely not a function
+                if (crumbs.length !== 7) {
+                    vscode.window.showErrorMessage("The parsed path doesn't lead to function.")
+                    return
+                }
+                // If the path doesn't contain required crumbs
+                if (!((crumbs[0] === "test.js" || crumbs[0] === "code.js") && crumbs[2] === "function" && crumbs[5] === "app")) {
+                    vscode.window.showErrorMessage("The parsed path doesn't correspond to the function test schema.")
+                    return
+                }
+                // If all checks passed, set URN
+                urn = `${_environment}/app/${crumbs[4]}/${crumbs[3]}/function`
+                functionName = `${crumbs[1]}`
             }
-            else if (context.name === "code" || context.name === "test") {
-                urn = `${_environment}/app/${Core.getApp(context).name}/${Core.getApp(context).version}/function/${context.parent.name}`
+
+            // Else parse from context
+            else {
+                // Set correct URN (if called from function or core or test)
+                urn = `${_environment}/app/${Core.getApp(context).name}/${Core.getApp(context).version}/function`
+                if (context.supertype === "function") {
+                    functionName = `${context.name}`
+                }
+                else if (context.name === "code" || context.name === "test") {
+                    functionName = `${context.parent.name}`
+                }
             }
 
             // Get current function code
-            let code = await Core.rpGet(`${urn}/code`, _authorization)
+            let code = await Core.rpGet(`${urn}/${functionName}/code`, _authorization)
 
             // Get current test code
-            let test = await Core.rpGet(`${urn}/test`, _authorization)
+            let test = await Core.rpGet(`${urn}/${functionName}/test`, _authorization)
+
+            // Get users' functions
+            let userFunctions = await Core.rpGet(`${urn}`, _authorization, { code: true })
 
             // Merge codes
             let codeToRun = `${code}\r\n\r\n/* === TEST CODE === */\r\n\r\n${test}`
@@ -72,30 +113,72 @@ class FunctionCommands {
              *  Sandbox cookbook
              *  - Assert for assertions
              *  - IML for internal IML functions
+             *  - Users' IML functions
              */
+            let success = 0
+            let fail = 0
+            let total = 0
             let sandbox = {
-                assert: require('assert'), iml: {}/*, it: (name, test) => {
+                assert: require('assert'),
+                iml: {},
+                it: (name, test) => {
+                    total++
                     outputChannel.append(`- ${name} ... `)
-                }*/
+                    try {
+                        test()
+                        outputChannel.appendLine(`✔`)
+                        success++
+                    }
+                    catch (err) {
+                        outputChannel.appendLine(`✘ => ${err}`)
+                        fail++
+                    }
+                },
+                environment: {
+                    timezone: _timezone
+                }
             };
 
             Object.keys(IML.FUNCTIONS).forEach(name => {
-                sandbox.iml[name] = IML.FUNCTIONS[name].value;
+                sandbox.iml[name] = IML.FUNCTIONS[name].value.bind({ timezone: sandbox.environment.timezone });
             });
 
-            // Prepare VM2, set timeout and pass assert
+            outputChannel.clear()
+            outputChannel.appendLine(`======= STARTING IML TEST =======\r\nFunction: ${functionName}\r\n---------- IN PROGRESS ----------`)
+
+            // Prepare VM2, set timeout and pass sandbox
             const vm = new VM({
                 timeout: 5000,
                 sandbox: sandbox
             })
 
+            vm.prepare = vm.run('(function(args) { global.__arguments__ = args })');
+
+            userFunctions.forEach(func => {
+                const preCompiled = new VMScript(`(${func.code}).apply({timezone: environment.timezone}, __arguments__)`, func.name);
+                sandbox.iml[func.name] = (...args) => {
+                    vm.prepare(args);
+                    return vm.run(preCompiled);
+                };
+            });
+
             // Show output channel IML tests and try running the test code
             outputChannel.show()
             try {
                 vm.run(codeToRun)
-                outputChannel.appendLine("Test OK")
+                outputChannel.appendLine(`----------- COMPLETED -----------`)
+                outputChannel.appendLine(`Total test blocks: ${total}`)
+                outputChannel.appendLine(`Passed blocks: ${success}`)
+                outputChannel.appendLine(`Failed blocks: ${fail}`)
+                if (fail === 0) {
+                    outputChannel.appendLine(`========== TEST PASSED ==========`)
+                }
+                else {
+                    outputChannel.appendLine(`========== TEST FAILED ==========`)
+                }
             }
             catch (err) {
+                outputChannel.appendLine(`========= CRITICAL FAIL =========`)
                 outputChannel.appendLine(err)
             }
         })