+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 1.0.4
+
+### Added
+
+- it function support in IML tests
+- Integromat Help command
+- Publish / Hide App or Module
+- IML tests with users' functions
+- Keybind and command for running IML test
+
+### Fixed
+
+- OpenSource apps icons
+- Webhook connection picker placeholder
+
+### Modified
+
+- Test output channel
+
 ## 1.0.3
 
 ### Added